LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/coop-navigated-history-popup.https-expected.txt

CONSOLE MESSAGE: Can't close the window since it was not opened by JavaScript

PASS Popup navigating to other origin with COOP: same-origin and back in history
